Migration/abstract notification service (#140)
* Refactoring: Abstract notification send service using interface

* Refactoring: Abstract notification history service using interface

* Refactoring: Abstract notification token manage service using interface
